Refactor kirby-carousel to replace the current kirby-slides component
