+++ apps/cookbook/src/app/examples/examples.common.ts
@@ -56,7 +56,6 @@ import { HeaderWithEmphasizedActionGroupExampleComponent } from './header-exampl
 import { HeaderWithCustomActionsExampleComponent } from './header-example/examples/custom-actions';
 import { HeaderWithInteractiveTitleExampleComponent } from './header-example/examples/interactive-title';
 import { MenuExampleComponent } from './menu-example/menu-example.component';
-import { CarouselExampleComponent } from './carousel-example/carousel-example.component';
 
 export const COMPONENT_DECLARATIONS: any[] = [
   ExamplesComponent,
@@ -118,5 +117,4 @@ export const COMPONENT_DECLARATIONS: any[] = [
   ListExperimentalExampleComponent,
   DataTableExampleComponent,
   MenuExampleComponent,
-  CarouselExampleComponent,
 ];
