+++ apps/cookbook/src/app/examples/slides-example/slides-example.component.scss
