+++ apps/cookbook/src/app/examples/slides-example/slides-example.component.ts
@@ -1,95 +1,91 @@
-import { Component, ViewChild } from '@angular/core';
-
-import { SlidesComponent } from '@kirbydesign/designsystem/slide';
-
-const config = {
-  selector: 'cookbook-slides-example',
-  template: `<kirby-slides
-  [slidesOptions]="slidesOptions"
-  [slides]="slides"
-  (selectedSlide)="getDataFromActiveSlide($event)"
->
-      <kirby-card  *kirbySlide="let slide; let i = index" [hasPadding]="true">
-        <kirby-card-header
-          [title]="slide.title"
-          [subtitle]="slide.subtitle"
-        ></kirby-card-header>
-
-        <!-- Card content example: -->
-        <div class="card-content">
-          {{ slide.cardContent }}
-        </div>
-    </kirby-card>
-</kirby-slides>
-<br />
-<div style="text-align: center">
-  <button kirby-button (click)="changeSlide()">Activate slide no. 4</button>
-</div>
-`,
-  codeSnippet: `slidesOptions = {
-  spaceBetween: 5,
-  centeredSlides: true,
-  slidesPerView: 1.15,
-  breakpoints: {
-    721: {
-      slidesPerView: 1.9,
-    },
-  },
-};
-
-slides = [
-  {
-    title: 'Slide 1',
-    subtitle: 'Subtitle 1',
-    cardContent:
-      ' Lorem, ipsum dolor ...',
-  },
-  {
-    title: 'Slide 2',
-    subtitle: 'Subtitle 2',
-    cardContent:
-    ' Lorem, ipsum dolor ...',
-  },
-  {
-    title: 'Slide 3',
-    subtitle: 'Subtitle 3',
-    cardContent:
-    ' Lorem, ipsum dolor ...',
-  },
-  {
-    title: 'Slide 4',
-    subtitle: 'Subtitle 4',
-    cardContent:
-    ' Lorem, ipsum dolor ...',
-  },
-  {
-    title: 'Slide 5',
-    subtitle: 'Subtitle 5',
-    cardContent:
-    ' Lorem, ipsum dolor ...',
-  },
-];
-`,
-};
+import { Component } from '@angular/core';
+import { KirbySwiperOptions } from '@kirbydesign/designsystem/slide';
+import { noop } from 'rxjs';
 
 @Component({
-  selector: config.selector,
-  template: config.template,
+  selector: 'cookbook-slides-example',
+  templateUrl: './slides-example.component.html',
+  styleUrls: ['./slides-example.component.scss'],
 })
 export class SlidesExampleComponent {
-  @ViewChild(SlidesComponent) slidesComponent: SlidesComponent;
-  template: string = config.template;
-  codeSnippet: string = config.codeSnippet;
+  maxWidth: string = 'full';
 
-  constructor() {}
+  maxWidthOptions = [
+    {
+      text: 'default',
+      value: 'default',
+    },
+    {
+      text: 'standard',
+      value: 'standard',
+    },
+    {
+      text: 'optimized',
+      value: 'optimized',
+    },
+    {
+      text: 'full',
+      value: 'full',
+    },
+  ];
 
-  slidesOptions = {
-    spaceBetween: 5,
+  ignorePagePadding: boolean = false;
+
+  onCheckedChange($event) {
+    this.ignorePagePadding = $event;
+  }
+
+  config: KirbySwiperOptions = {
+    slidesPerView: 1.2,
     centeredSlides: true,
-    slidesPerView: 1.15,
+    slidesPerGroup: 1,
     breakpoints: {
-      721: {
-        slidesPerView: 1.9,
+      992: {
+        centeredSlides: false,
+        slidesPerView: 3,
+        slidesPerGroup: 3,
+      },
+    },
+  };
+
+  config1: KirbySwiperOptions = {
+    slidesPerView: 1.2,
+    centeredSlides: true,
+    slidesPerGroup: 1,
+    speed: 400,
+    breakpoints: {
+      992: {
+        centeredSlides: false,
+        slidesPerView: 3,
+        slidesPerGroup: 3,
+      },
+    },
+  };
+
+  config2: KirbySwiperOptions = {
+    slidesPerView: 1.2,
+    centeredSlides: true,
+    slidesPerGroup: 1,
+    speed: 500,
+    breakpoints: {
+      992: {
+        centeredSlides: false,
+        slidesPerView: 3,
+        slidesPerGroup: 3,
+      },
+    },
+  };
+
+  config3: KirbySwiperOptions = {
+    slidesPerView: 1.2,
+    centeredSlides: true,
+    slidesPerGroup: 1,
+    speed: 600,
+    breakpoints: {
+      992: {
+        centeredSlides: false,
+        slidesPerView: 3,
+        slidesPerGroup: 3,
       },
     },
   };
@@ -99,45 +95,59 @@ export class SlidesExampleComponent {
       title: 'Slide 1',
       subtitle: 'Subtitle 1',
       cardContent:
-        ' Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat!',
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
     },
     {
       title: 'Slide 2',
       subtitle: 'Subtitle 2',
       cardContent:
-        ' Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat!',
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
     },
     {
       title: 'Slide 3',
       subtitle: 'Subtitle 3',
       cardContent:
-        ' Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat!',
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
     },
     {
       title: 'Slide 4',
       subtitle: 'Subtitle 4',
       cardContent:
-        ' Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat!',
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
     },
     {
       title: 'Slide 5',
       subtitle: 'Subtitle 5',
       cardContent:
-        ' Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat!',
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
+    },
+    {
+      title: 'Slide 6',
+      subtitle: 'Subtitle 6',
+      cardContent:
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
+    },
+    {
+      title: 'Slide 7',
+      subtitle: 'Subtitle 7',
+      cardContent:
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
+    },
+    {
+      title: 'Slide 8',
+      subtitle: 'Subtitle 8',
+      cardContent:
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
+    },
+    {
+      title: 'Slide 9',
+      subtitle: 'Subtitle 9',
+      cardContent:
+        '  Lorem, ipsum dolor sit amet consectetur adipisicing elit. Mollitia facere molestias recusandae necessitatibus ab veniam repellendus doloremque culpa quam libero, est quo  accusamus cumque, in quia itaque cupiditate ratione repellat! ',
     },
   ];
 
-  data: any;
-  selectedSlideIndex: number;
-
-  getDataFromActiveSlide(activeSlide: { selectedData: any; selectedSlideIndex: number }) {
-    this.data = activeSlide.selectedData;
-    this.selectedSlideIndex = activeSlide.selectedSlideIndex;
-
-    console.log('Output: ', activeSlide);
-  }
-
-  changeSlide() {
-    this.slidesComponent.slideTo(3);
+  getDataFromActiveSlide($event) {
+    console.log($event);
   }
 }
