--- apps/cookbook/src/app/showcase/carousel-showcase/carousel-showcase.component.html
@@ -1,72 +1,0 @@
-<p>
-  Carousels are a collection of cards you can navigate horizontally in, either swipe or click.
-  Carousels are great to contain content without using a lot of space, but the content has to be
-  closely related.
-</p>
-
-<h2>Example:</h2>
-
-<cookbook-iphone
-  viewMode="full-size"
-  [showViewModeToggle]="true"
-  src="/examples/carousel/"
-  showExternalLink="true"
-></cookbook-iphone>
-
-<cookbook-code-viewer [html]="exampleComponentHTML"></cookbook-code-viewer>
-
-<p>
-  The
-  <code>kirby-carousel</code>
-  uses
-  <code>swiper.js</code>
-  under the hood and accepts the same
-  <a href="https://swiperjs.com/swiper-api#parameters" target="_blank">configuration</a>
-  as
-  <code>swiper.js</code>
-  . The component comes with this default config:
-</p>
-
-<cookbook-example-viewer [ts]="carouselDefaultConfig"></cookbook-example-viewer>
-
-<br />
-
-<p>
-  This config can be extended be passing in
-  <code>slidesOptions</code>
-  as shown in the example above, which uses this config:
-</p>
-
-<cookbook-code-viewer [ts]="customConfigExample"></cookbook-code-viewer>
-
-<p>
-  <em>
-    <strong>Please note:</strong>
-    When providing your own config it is important that you don't specify
-    <code>pagination</code>
-    ,
-    <code>navigation</code>
-    &
-    <code>slideChange</code>
-    , as these properties will be overriden by the default config to ensure the component works as
-    expected.
-  </em>
-</p>
-
-<h2>Slots</h2>
-
-<p>
-  The component has a default slot, which will place a button in the top right corner next to the
-  pagination and navigation elements.
-</p>
-
-<h2 #fullscreenModalMethods>Properties:</h2>
-<cookbook-api-description-properties
-  [properties]="properties"
-></cookbook-api-description-properties>
-
-<h2 #fullscreenModalMethods>Events:</h2>
-<cookbook-api-description-events [events]="events"></cookbook-api-description-events>
-
-<h2 #fullscreenModalMethods>Methods:</h2>
-<cookbook-api-description-methods [methods]="methods"></cookbook-api-description-methods>
