+++ apps/cookbook/src/app/showcase/slides-showcase/slides-showcase.component.ts
@@ -1,39 +1,61 @@
 import { Component } from '@angular/core';
-import { ApiDescriptionEvent } from '~/app/shared/api-description/api-description-events/api-description-events.component';
-import { ApiDescriptionProperty } from '~/app/shared/api-description/api-description-properties/api-description-properties.component';
-
+import {
+  customConfigExample,
+  exampleComponentHTML,
+  slidesDefaultConfig,
+} from '../../examples/slides-example/slides-code-snippets';
+import { ApiDescriptionProperty } from '../../shared/api-description/api-description-properties/api-description-properties.component';
+import { ApiDescriptionEvent } from '../../shared/api-description/api-description-events/api-description-events.component';
+import { ApiDescriptionMethod } from '../../shared/api-description/api-description-methods/api-description-methods.component';
 @Component({
-  selector: 'cookbook-slide-showcase',
+  selector: 'cookbook-carousel-showcase',
   templateUrl: './slides-showcase.component.html',
-  preserveWhitespaces: true,
 })
 export class SlidesShowcaseComponent {
+  slidesDefaultConfig: string = slidesDefaultConfig;
+  exampleComponentHTML: string = exampleComponentHTML;
+  customConfigExample: string = customConfigExample;
+
   properties: ApiDescriptionProperty[] = [
     {
       name: 'slidesOptions',
-      description:
-        'Configuration object for ion-slides. View documentation at: [https://ionicframework.com/docs/v3/api/components/slides/Slides/#input-properties] for available props',
-      defaultValue: '',
-      type: ['object'],
+      description: `Configuration object for swiper-container. View documentation at: [https://swiperjs.com/swiper-api#parameters] for available parameters`,
+      defaultValue: 'null',
+      type: ['KirbySwiperOptions'],
     },
     {
       name: 'slides',
-      description: 'Object array holding the data to be shown',
-      defaultValue: '',
-      type: ['DataArray'],
+      description: `Array of items to be displayed in the carousel.`,
+      defaultValue: 'null',
+      type: ['any[]'],
     },
     {
-      name: 'activeSlideIndex',
-      description: 'The index of the active slide',
-      defaultValue: '0',
-      type: ['number'],
+      name: 'title',
+      description: `A title that will be displayed above the carousel in the top left corner.`,
+      defaultValue: 'null',
+      type: ['string'],
+    },
+    {
+      name: 'ignorePagePadding',
+      description: `Makes the carousel ignore the kirby-page padding and expand to the edge of the page.`,
+      defaultValue: 'null',
+      type: ['boolean'],
     },
   ];
+
   events: ApiDescriptionEvent[] = [
     {
       name: 'selectedSlide',
-      description: 'Emits active slide´s data and index',
-      signature: 'func',
+      description: 'Emits the active slide´s data and index',
+      signature: '{slide: any, index: number}',
+    },
+  ];
+
+  methods: ApiDescriptionMethod[] = [
+    {
+      name: 'slideTo',
+      description: 'Slides to the specified slide index.',
+      signature: '(index: number) => void',
     },
   ];
 }
