+++ apps/flows/tsconfig.json
@@ -81,8 +81,7 @@
         "libs/designsystem/alert-experimental/index.ts"
       ],
       "@kirbydesign/designsystem/action-group": ["libs/designsystem/action-group/index.ts"],
-      "@kirbydesign/designsystem/menu": ["libs/designsystem/menu/index.ts"],
-      "@kirbydesign/designsystem/carousel": ["libs/designsystem/carousel/index.ts"]
+      "@kirbydesign/designsystem/menu": ["libs/designsystem/menu/index.ts"]
     },
     "target": "es2020"
   },
