--- libs/designsystem/carousel/index.ts
@@ -1,1 +1,0 @@
-export * from './src/index';
