--- libs/designsystem/carousel/ng-package.json
@@ -1,1 +1,0 @@
-{}
