--- libs/designsystem/carousel/src/carousel-slide.directive.ts
@@ -1,6 +1,0 @@
-import { Directive } from '@angular/core';
-
-@Directive({
-  selector: '[kirbyCarouselSlide]',
-})
-export class CarouselSlideDirective {}
