--- libs/designsystem/carousel/src/index.ts
@@ -1,1 +1,0 @@
-export * from './public_api';
