--- libs/designsystem/carousel/src/public_api.ts
@@ -1,3 +1,0 @@
-export { CarouselComponent, KirbySwiperOptions } from './carousel.component';
-export { CarouselSlideDirective } from './carousel-slide.directive';
-export { CarouselModule } from './carousel.module';
