+++ libs/designsystem/slide/src/slide.module.ts
@@ -1,14 +1,17 @@
 import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { KirbyIonicModule } from '@kirbydesign/designsystem/kirby-ionic-module';
-
-import { SlideDirective } from './slide.directive';
+import { ButtonComponent } from '@kirbydesign/designsystem/button';
+import { IconModule } from '@kirbydesign/designsystem/icon';
 import { SlidesComponent } from './slides.component';
+import { SlideDirective } from './slide.directive';
 
-const declarations = [SlideDirective, SlidesComponent];
+const declarations = [SlidesComponent, SlideDirective];
+
 @NgModule({
+  imports: [CommonModule, KirbyIonicModule, ButtonComponent, IconModule],
   declarations: [...declarations],
-  imports: [KirbyIonicModule, CommonModule],
   exports: [...declarations],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
 export class SlideModule {}
