+++ libs/designsystem/slide/src/slides.component.scss
