+++ libs/designsystem/tsconfig.spec.json
@@ -58,8 +58,7 @@
       "@kirbydesign/designsystem/toast": ["toast/index.ts"],
       "@kirbydesign/designsystem/alert-experimental": ["alert-experimental/index.ts"],
       "@kirbydesign/designsystem/action-group": ["action-group/index.ts"],
-      "@kirbydesign/designsystem/menu": ["menu/index.ts"],
-      "@kirbydesign/designsystem/carousel": ["carousel/index.ts"]
+      "@kirbydesign/designsystem/menu": ["menu/index.ts"]
     }
   },
   "files": ["test.ts"],
